Extracted MessageLog class to a separate file
We don't want to co-mingle these classes as it really makes it hard to
maintain the codebase.
